fix type conversion warnings on new dashboard machine
fixing dashboard issues

+++ IO/Core/vtkLZ4DataCompressor.cxx
@@ -47,7 +47,10 @@ vtkLZ4DataCompressor::CompressBuffer(unsigned char const* uncompressedData,
   const char *ud = reinterpret_cast<const char *>(uncompressedData);
   char *cd = reinterpret_cast<char*>(compressedData);
   // Call LZ4's compress function.
-  int cs = LZ4_compress_fast(ud, cd, uncompressedSize, compressionSpace, this->AccelerationLevel);
+  int cs =
+    LZ4_compress_fast(ud, cd,
+      static_cast<int>(uncompressedSize),
+      static_cast<int>(compressionSpace), this->AccelerationLevel);
   if (cs == 0)
   {
     vtkErrorMacro("LZ4 error while compressing data.");
@@ -64,8 +67,11 @@ vtkLZ4DataCompressor::UncompressBuffer(unsigned char const* compressedData,
 {
   char* ud = reinterpret_cast<char*>(uncompressedData);
   const char* cd = reinterpret_cast<const char*>(compressedData);
-  int us = LZ4_decompress_safe(cd,ud,compressedSize,uncompressedSize);
-  if(us < 0)
+  int us =
+    LZ4_decompress_safe(cd,ud,
+      static_cast<int>(compressedSize),
+      static_cast<int>(uncompressedSize));
+  if (us < 0)
   {
     vtkErrorMacro("Zlib error while uncompressing data.");
     return 0;
